+++ CHANGELOG.md
@@ -5,13 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [*3.1.0*] - <*2021-03-01*>
+
+### Added
+
+* Added `setThemeHandler` option (#13)
+* Added unit testing
+
+### Fixes
+
+* Fixed an issue with the `@fluentui/react-northstar` peer dependency version range with npm 7+
+
 ## [*3.0.2*] - <*2021-01-29*>
 
-### Changes
+### Fixes
 
 * Fixed issue where the Teams JS Library was not initialized while hosted outside of Teams (for instance in a popup)
 
-## [*3.0.1*] - <*2012-12-18*>
+## [*3.0.1*] - <*2020-12-18*>
 
 ### Changes